# swagger.yaml
openapi: 3.0.3
info:
  title: Elasticsearch PoC API
  description: |
    A comprehensive Proof of Concept API for evaluating Elasticsearch as a search and analytics engine.
    
    This API provides endpoints for:
    - Product management (CRUD operations)
    - Advanced search capabilities with multiple query types
    - Analytics and aggregations
    - Performance testing and monitoring
    
    ## Features Tested
    - Full-text search with relevance scoring
    - Multi-field search across name, description, category
    - Fuzzy search with auto-correction
    - Wildcard and regex pattern matching
    - Boolean queries with must/should/filter clauses
    - Range filters for price, rating, dates
    - Faceted search with aggregation-based filters
    - Statistical and bucket aggregations
    - Performance benchmarking and load testing
    
  version: 1.0.0
  contact:
    name: Lawsikho Team
    email: support@lawsikho.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server (example)
...
paths:
...
  /api/analytics/overview:
    get:
      tags:
        - Analytics
      summary: Get overview analytics
      description: Get high-level analytics and statistics
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytics/categories:
    get:
      tags:
        - Analytics
      summary: Get category distribution
      description: Get analytics for product categories
      parameters:
        - name: size
          in: query
          description: Number of categories to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Category analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytics/price-distribution:
    get:
      tags:
        - Analytics
      summary: Get price distribution
      description: Get analytics for product price distribution
      responses:
        '200':
          description: Price analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytics/rating-distribution:
    get:
      tags:
        - Analytics
      summary: Get rating distribution
      description: Get analytics for product rating distribution
      responses:
        '200':
          description: Rating analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytics/trends:
    get:
      tags:
        - Analytics
      summary: Get time-based trends
      description: Get analytics for time-based trends
      parameters:
        - name: interval
          in: query
          description: Time interval for trends
          schema:
            type: string
            enum: [day, week, month]
            default: day
        - name: size
          in: query
          description: Number of time periods to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: Trends retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytics/search-analytics:
    get:
      tags:
        - Analytics
      summary: Get search analytics
      description: Get analytics for search performance
      parameters:
        - name: days
          in: query
          description: Number of days to analyze
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
      responses:
        '200':
          description: Search analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAnalytics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    OverviewAnalytics:
      type: object
      properties:
        total_products:
          type: integer
          description: Total number of products
          example: 1000
        price_stats:
          type: object
          properties:
            average:
              type: number
              format: double
              example: 299.99
            min:
              type: number
              format: double
              example: 10.00
            max:
              type: number
              format: double
              example: 999.99
        average_rating:
          type: number
          format: double
          example: 4.2
        stock_status:
          type: object
          properties:
            in_stock:
              type: integer
              example: 850
            out_of_stock:
              type: integer
              example: 150
    CategoryAnalytics:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
                example: "Electronics"
              count:
                type: integer
                example: 150
              avg_price:
                type: number
                format: double
                example: 299.99
              avg_rating:
                type: number
                format: double
                example: 4.2
              in_stock_count:
                type: integer
                example: 120
              in_stock_percentage:
                type: integer
                example: 80
    PriceAnalytics:
      type: object
      properties:
        price_ranges:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
                example: "$100 - $200"
              count:
                type: integer
                example: 75
              avg_rating:
                type: number
                format: double
                example: 4.1
              in_stock_count:
                type: integer
                example: 60
        histogram:
          type: array
          items:
            type: object
            properties:
              price:
                type: number
                format: double
                example: 150
              count:
                type: integer
                example: 25
    RatingAnalytics:
      type: object
      properties:
        rating_ranges:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
                example: "4 - 5"
              count:
                type: integer
                example: 100
              avg_price:
                type: number
                format: double
                example: 350.00
              in_stock_count:
                type: integer
                example: 85
        top_rated:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Premium Smartphone"
              count:
                type: integer
                example: 5
              avg_rating:
                type: number
                format: double
                example: 4.8
              avg_price:
                type: number
                format: double
                example: 599.99
    TrendsAnalytics:
      type: object
      properties:
        trends:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
                example: "2024-01-01"
              timestamp:
                type: integer
                example: 1704067200000
              count:
                type: integer
                example: 10
              avg_price:
                type: number
                format: double
                example: 299.99
              avg_rating:
                type: number
                format: double
                example: 4.2
              top_categories:
                type: array
                items:
                  type: object
                  properties:
                    category:
                      type: string
                      example: "Electronics"
                    count:
                      type: integer
                      example: 5
    SearchAnalytics:
      type: object
      properties:
        period_days:
          type: integer
          example: 7
        total_searches:
          type: integer
          example: 500
        top_search_terms:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                example: "smartphone"
              doc_count:
                type: integer
                example: 50
        categories_searched:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                example: "Electronics"
              doc_count:
                type: integer
                example: 200
        price_ranges_searched:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                example: "100-200"
              doc_count:
                type: integer
                example: 150
...
    Error:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Error message
          example: "Internal Server Error"
        message:
          type: string
          description: Detailed error message
          example: "Something went wrong"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00.000Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []